# control-api/docs/swagger.yaml
basePath: /
definitions:
  auth.authResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/auth.userResponse'
    type: object
  auth.changePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  auth.errorResponse:
    properties:
      error:
        type: string
    type: object
  auth.forgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth.forgotPasswordResponse:
    properties:
      message:
        type: string
      reset_url:
        description: ResetURL is returned temporarily until SMTP delivery is wired
          up.
        type: string
    type: object
  auth.googleAuthRequest:
    properties:
      id_token:
        type: string
    required:
    - id_token
    type: object
  auth.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.registerRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  auth.resetPasswordRequest:
    properties:
      password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  auth.userResponse:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
host: localhost:9090
info:
  contact: {}
  description: Control plane for xpaywall — manages projects, routes, users, payment
    channels, and request logs.
  title: xpaywall Control API
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.forgotPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.errorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.forgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.forgotPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.errorResponse'
      summary: Request password reset
      tags:
      - auth
  /auth/google:
    post:
      consumes:
      - application/json
      parameters:
      - description: Google ID token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.googleAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.errorResponse'
      summary: Google sign-in
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.errorResponse'
      summary: Login
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.userResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth.errorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.registerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.errorResponse'
      summary: Register
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.forgotPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.errorResponse'
      summary: Reset password
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    description: Internal API key shared between xgateway and control-api
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: 'JWT token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"